Refactor Poincaré map configuration and enhance computation methods
- Renamed PoincareMapConfig to poincareMapConfig for consistency with naming conventions.
- Updated main.py to utilize the new configuration class and improved initialization parameters.
- Enhanced Poincaré map generation by integrating GPU support and refining the seed generation process.
- Streamlined import statements and removed unused functions for better code clarity.
- Improved error handling and logging during Poincaré map computations.

+++ src/algorithms/center/polynomial/algebra.py
@@ -416,7 +416,7 @@ def _poly_evaluate(
     if p.shape[0] == 0:
         return 0.0 + 0.0j
 
-    pow_table = np.empty((N_VARS, degree + 1), dtype=point.dtype)
+    pow_table = np.empty((N_VARS, degree + 1), dtype=np.complex128)
     for v in range(N_VARS):
         pow_table[v, 0] = 1.0 + 0.0j
         base = point[v]
